+++ swirly-fig/src/main/java/com/swirlycloud/engine/Market.java
@@ -165,9 +165,9 @@ public final class Market extends BasicRbNode implements Identifiable, Printable
 
     public static long composeId(long contrId, int settlDay) {
         // 16 bit contr-id.
-        final int CONTR_MASK = (1 << 16) - 1;
+        final long CONTR_MASK = (1L << 16) - 1;
         // 16 bits is sufficient for truncated Julian day.
-        final int TJD_MASK = (1 << 16) - 1;
+        final long TJD_MASK = (1L << 16) - 1;
 
         // Truncated Julian Day (TJD).
         final long tjd = Date.jdToTjd(settlDay);
